backend/requirements/base.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-decouple==3.8
celery==5.3.6
django-celery-beat==2.5.0
redis==5.0.1
stripe==8.2.0
cloudinary==1.38.0
resend==0.7.0
sentry-sdk==1.40.0
Pillow==10.2.0
django-ratelimit==4.1.0
social-auth-app-django==5.4.0
